requirements.txt
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
path==16.14.0
PyYAML==6.0.1
